// src/contexts/types.ts
// Copyright 2024 Tether Operations Limited
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

import type {
  AddressMap,
  Amount,
  BalanceMap,
  ChainsConfig,
  Transaction,
  TransactionMap,
  Wallet,
} from '../services/wdk-service/types';

export interface WalletContextState {
  wallet?: Wallet | null;
  addresses?: AddressMap;
  balances: {
    list: Amount[];
    map: BalanceMap;
    isLoading: boolean;
  };
  transactions: {
    list: Transaction[];
    map: TransactionMap;
    isLoading: boolean;
  };
  isLoading: boolean;
  error: string | null;
  isInitialized: boolean;
  isUnlocked: boolean;
}

export interface WalletContextType extends WalletContextState {
  createWallet: (params: {
    name: string;
    mnemonic?: string;
  }) => Promise<WalletContextState['wallet']>;
  clearWallet: () => Promise<void>;
  clearError: () => void;
  refreshWalletBalance: () => Promise<void>;
  refreshTransactions: () => Promise<void>;
  unlockWallet: () => Promise<boolean | undefined>;
}

// Provider configuration
export interface WalletProviderConfig {
  indexer: {
    apiKey: string;
    url: string;
    version?: string;
  };
  chains: ChainsConfig;
  enableCaching?: boolean;
}